+++ wxcadm/virtual_line.py
@@ -167,6 +167,44 @@ class VirtualLine:
                                   payload=self._build_put_payload())
         return True
 
+    @property
+    def ecbn(self) -> dict:
+        """ The Emergency Callback Number details of the Virtual Line """
+        response = webex_api_call('get', f'v1/telephony/config/virtualLines/{self.id}/emergencyCallbackNumber',
+                                  params={'orgId': self.parent.org_id})
+        return response
+
+    def set_ecbn(self, value: Union[str, wxcadm.Person, wxcadm.Workspace, wxcadm.VirtualLine]):
+        """ Set the ECBN of the Virtual Line
+
+        Valid values are ``'direct'``, ``'location'``, or a :class:`Person`, :class:`Workspace`, or
+        :class:`VirtualLine` to set the ECBN to one of those.
+
+        Args:
+            value (str, Person, Workspace, VirtualLine): The value to set the ECBN to
+
+        Returns:
+            bool: True on success
+
+        """
+        if isinstance(value, wxcadm.Person) or \
+                isinstance(value, wxcadm.Workspace) or \
+                isinstance(value, wxcadm.VirtualLine):
+            payload = {
+                'selected': 'LOCATION_MEMBER_NUMBER',
+                'locationMemberId': value.id
+            }
+        elif value.lower() == 'direct' or value.lower() == 'direct_line':
+            payload = {'selected': 'DIRECT_LINE'}
+        elif value.lower() == 'location' or value.lower() == 'location_ecbn':
+            payload = {'selected': 'LOCATION_ECBN'}
+        else:
+            raise ValueError('Unknown value')
+
+        response = webex_api_call('put', f'v1/telephony/config/virtualLines/{self.id}/emergencyCallbackNumber',
+                                  params={'orgId': self.org_id}, payload=payload)
+        return response
+
 
 class VirtualLineList(UserList):
     _endpoint = "v1/telephony/config/virtualLines"
@@ -217,16 +255,19 @@ class VirtualLineList(UserList):
     def get(self,
             id: Optional[str] = None,
             first_name: Optional[str] = None,
-            last_name: Optional[str] = None) -> Optional[VirtualLine]:
+            last_name: Optional[str] = None,
+            name: Optional[str] = None) -> Optional[VirtualLine]:
         """ Get the instance associated with a given ID or name.
 
         When searching by name, both `first_name` and `last_name` are required. Webex does not have a single key value
-        other than ID, but the combination of first and last name must be unique.
+        other than ID, but the combination of first and last name must be unique. For names with only a single first
+        name and single last name, such as 'Joe Smith', the ``name`` argument can be used as a shortcut.
 
         Args:
             id (str, optional): The Virtual Line ID to find
             first_name (str, optional): The first name of the Virtual Line
             last_name (str, optional): The last name of the Virtual Line
+            name (str, optional): The full name (first and last) of the Virtual Line
 
         Returns:
             VirtualLine: The VirtualLine instance correlating to the given search argument. None is returned if no
@@ -240,7 +281,9 @@ class VirtualLineList(UserList):
             for item in self.data:
                 if item.id == id:
                     return item
-        if first_name is not None or last_name is not None:
+        if first_name is not None or last_name is not None or name is not None:
+            if name is not None and first_name is None and last_name is None:
+                first_name, last_name = name.split(' ', 1)
             if first_name and last_name:
                 for item in self.data:
                     if item.first_name.lower() == first_name.lower() and item.last_name.lower() == last_name.lower():