+++ wxcadm/workspace.py
@@ -439,6 +439,44 @@ class Workspace:
         self.calendar = config.get('calendar', None)
         self.notes = config.get("notes", "")
 
+    @property
+    def ecbn(self) -> dict:
+        """ The Emergency Callback Number details of the Workspace """
+        response = webex_api_call('get', f'v1/telephony/config/workspaces/{self.id}/emergencyCallbackNumber',
+                                  params={'orgId': self.org_id})
+        return response
+
+    def set_ecbn(self, value: Union[str, wxcadm.Person, wxcadm.Workspace, wxcadm.VirtualLine]):
+        """ Set the ECBN of the Workspace
+
+        Valid values are ``'direct'``, ``'location'``, or a :class:`Person`, :class:`Workspace`, or
+        :class:`VirtualLine` to set the ECBN to one of those.
+
+        Args:
+            value (str, Person, Workspace, VirtualLine): The value to set the ECBN to
+
+        Returns:
+            bool: True on success
+
+        """
+        if isinstance(value, wxcadm.Person) or \
+                isinstance(value, wxcadm.Workspace) or \
+                isinstance(value, wxcadm.VirtualLine):
+            payload = {
+                'selected': 'LOCATION_MEMBER_NUMBER',
+                'locationMemberId': value.id
+            }
+        elif value.lower() == 'direct' or value.lower() == 'direct_line':
+            payload = {'selected': 'DIRECT_LINE'}
+        elif value.lower() == 'location' or value.lower() == 'location_ecbn':
+            payload = {'selected': 'LOCATION_ECBN'}
+        else:
+            raise ValueError('Unknown value')
+
+        response = webex_api_call('put', f'v1/telephony/config/workspaces/{self.id}/emergencyCallbackNumber',
+                                  params={'orgId': self.org_id}, payload=payload)
+        return response
+
 
 class WorkspaceLocation:
     def __init__(self, parent: wxcadm.Org, id: str, config: dict = None):